+++ frontend2/src/components/tables/chart/randomData.ts
